pyproject.toml
[project]
name = "dspy-query-wizard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.1",
    "dspy-ai>=3.0.4",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "joblib>=1.5.2",
    "openai>=1.0.0",
    "python-dotenv>=1.2.1",
    "transformers>=4.57.1",
    "typing-extensions>=4.12.2",
    "uvicorn>=0.38.0",
    "openai>=1.0.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.28.0",
    "pytest>=9.0.1",
    "pyright>=1.1.396",
    "ruff>=0.9.9",
    "pytest-asyncio>=1.3.0",
]

[tool.pyright]
# Only ignore missing stubs for libraries that truly don't have them
reportMissingTypeStubs = false 
